fix: use Ubuntu 18.04 for Vault image
Associating an IAM instance profile to an EC2 instance is supposed to
propagate almost immediately; however, this is not always true.

The EC2 instances require read and write access to the DynamoDB table,
but Vault failed to start claiming that it did not have access to the
table.

The version of systemd is believed to be the source of this issue.

Mainly, a race condition is occurring from the time the IAM instance
profile is associated with the EC2 instance to the time the Vault
service starts.

To resolve this issue, upgrade the Ubuntu distribution from 16.04 to
18.04. Ubuntu 18.04 has a newer version of systemd installed and allows
to specify a timeout between restarts of the service.